Removing unused line of code
This markdown comment was thought to be where the identifier string was being added. However, it actually is not returned from the go-patch-cover tool. The special markdown comment is added in the next step. This was refactored in the previous commit so that the same string format and values are used for the text and lookup.